[back] 유저ID 세션으로 가져오게 변경

+++ codearea/src/main/java/sample/codearea/service/AnswerService.java
@@ -1,8 +1,11 @@
 package sample.codearea.service;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import sample.codearea.constant.SessionConst;
 import sample.codearea.dto.AnswerRequestDto;
 import sample.codearea.dto.AnswerResponseDto;
 import sample.codearea.entity.AnswerEntity;
@@ -39,8 +42,9 @@ public class AnswerService {
                 .collect(Collectors.toList());
     }
 
-    public AnswerResponseDto save(Long userId, Long questionId, AnswerRequestDto answerRequestDto) {
-        UserEntity user = userRepository.findById(userId).get();
+    public AnswerResponseDto save(Long questionId, AnswerRequestDto answerRequestDto, HttpServletRequest httpServletRequest) {
+        Long loginId = getLoginId(httpServletRequest);
+        UserEntity user = userRepository.findById(loginId).get();
         QuestionEntity question = questionRepository.findById(questionId)
                 .orElseThrow(() -> new IllegalArgumentException("Not found Question"));
 
@@ -59,17 +63,33 @@ public class AnswerService {
         return answerResponseDto;
     }
 
-    public void update(Long answerId, AnswerRequestDto answerRequestDto) {
-        AnswerEntity answer = answerRepository.findById(answerId).orElseThrow();
+    public void update(Long answerId, AnswerRequestDto answerRequestDto, HttpServletRequest httpServletRequest) {
+        Long loginId = getLoginId(httpServletRequest);
+        AnswerEntity answer = answerRepository.findById(answerId).orElseThrow(() -> new IllegalArgumentException("not found answer"));
 
-        answer.setContent(answerRequestDto.getContent());
-        answerRepository.save(answer);
+        if(loginId == answer.getUser().getId()){
+            answer.setContent(answerRequestDto.getContent());
+            answerRepository.save(answer);
+            return;
+        }
+        throw new IllegalArgumentException("허가되지 않은 사용자입니다.");
     }
 
-    public void delete(Long answerId) {
-        AnswerEntity answer = answerRepository.findById(answerId).orElseThrow();
-        answerRepository.delete(answer);
+    public void delete(Long answerId, HttpServletRequest httpServletRequest) {
+        Long loginId = getLoginId(httpServletRequest);
+        AnswerEntity answer = answerRepository.findById(answerId).orElseThrow(() -> new IllegalArgumentException("not found answer"));
+
+        if(loginId == answer.getUser().getId()){
+            answerRepository.delete(answer);
+            return;
+        }
+        throw new IllegalArgumentException("허가되지 않은 사용자입니다.");
     }
 
+    private Long getLoginId(HttpServletRequest httpServletRequest) {
+        HttpSession session = httpServletRequest.getSession();
+        Long loginId = (Long) session.getAttribute(SessionConst.LOGIN_USER);
+        return loginId;
+    }
 
 }
